Set text block dialog colors to the same values as the viewer pane
The viewer pane and text block dialog iframes do not inherit CSS from
the parent document, but they do have transparent backgrounds. We hard
code the main viewer color and background-color values to black and
white respectively, but did not for the text block viewer iframe. This
caused the text block viewer pane to show black text on a black
background when dark mode was enabled.

This change uses the same hard coded color and background colors for the
text block viewer dialog as we use for the document viewer.